add store.js and move notifications
