+++ bloglist/frontend/src/main.jsx
@@ -2,19 +2,7 @@ import ReactDOM from 'react-dom/client'
 import App from './App'
 import './services/index.css'
 import { Provider } from 'react-redux'
-import { configureStore } from '@reduxjs/toolkit'
-
-import notificationReducer from './reducers/notificationReducer'
-import blogReducer from './reducers/blogReducer'
-import userReducer from './reducers/userReducer'
-
-const store = configureStore({
-  reducer: {
-    notification: notificationReducer,
-    blogs: blogReducer,
-    user: userReducer
-  }
-})
+import store from './reducers/store'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <Provider store={store}>
